requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
alembic==1.13.1
psycopg==3.1.18